global variable dependants returned to splash further experiments to be done
