--- Urho3D/splash_build/CMakeFiles/Progress/5
@@ -1,1 +1,0 @@
-empty
