--- Urho3D/splash_build/CMakeFiles/Progress/8
@@ -1,1 +1,0 @@
-empty
