+++ Urho3D/splash_build/CMakeFiles/Splash.dir/build.make
@@ -128,22 +128,9 @@ CMakeFiles/Splash.dir/playNote.cpp.s: cmake_force
 	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/Splash.dir/playNote.cpp.s"
 	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/pi/sound-pirates/Urho3D/splash/playNote.cpp -o CMakeFiles/Splash.dir/playNote.cpp.s
 
-CMakeFiles/Splash.dir/processBuffer.cpp.o: CMakeFiles/Splash.dir/flags.make
-CMakeFiles/Splash.dir/processBuffer.cpp.o: /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_5) "Building CXX object CMakeFiles/Splash.dir/processBuffer.cpp.o"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/Splash.dir/processBuffer.cpp.o -c /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp
-
-CMakeFiles/Splash.dir/processBuffer.cpp.i: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing CXX source to CMakeFiles/Splash.dir/processBuffer.cpp.i"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -E /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp > CMakeFiles/Splash.dir/processBuffer.cpp.i
-
-CMakeFiles/Splash.dir/processBuffer.cpp.s: cmake_force
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling CXX source to assembly CMakeFiles/Splash.dir/processBuffer.cpp.s"
-	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -S /home/pi/sound-pirates/Urho3D/splash/processBuffer.cpp -o CMakeFiles/Splash.dir/processBuffer.cpp.s
-
 CMakeFiles/Splash.dir/realtime-beat.cpp.o: CMakeFiles/Splash.dir/flags.make
 CMakeFiles/Splash.dir/realtime-beat.cpp.o: /home/pi/sound-pirates/Urho3D/splash/realtime-beat.cpp
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_6) "Building CXX object CMakeFiles/Splash.dir/realtime-beat.cpp.o"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_5) "Building CXX object CMakeFiles/Splash.dir/realtime-beat.cpp.o"
 	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/Splash.dir/realtime-beat.cpp.o -c /home/pi/sound-pirates/Urho3D/splash/realtime-beat.cpp
 
 CMakeFiles/Splash.dir/realtime-beat.cpp.i: cmake_force
@@ -156,7 +143,7 @@ CMakeFiles/Splash.dir/realtime-beat.cpp.s: cmake_force
 
 CMakeFiles/Splash.dir/splash.cpp.o: CMakeFiles/Splash.dir/flags.make
 CMakeFiles/Splash.dir/splash.cpp.o: /home/pi/sound-pirates/Urho3D/splash/splash.cpp
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_7) "Building CXX object CMakeFiles/Splash.dir/splash.cpp.o"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_6) "Building CXX object CMakeFiles/Splash.dir/splash.cpp.o"
 	/usr/bin/c++ $(CXX_DEFINES) $(CXX_INCLUDES) $(CXX_FLAGS) -o CMakeFiles/Splash.dir/splash.cpp.o -c /home/pi/sound-pirates/Urho3D/splash/splash.cpp
 
 CMakeFiles/Splash.dir/splash.cpp.i: cmake_force
@@ -173,7 +160,6 @@ Splash_OBJECTS = \
 "CMakeFiles/Splash.dir/fft.cpp.o" \
 "CMakeFiles/Splash.dir/instructionsStatements.cpp.o" \
 "CMakeFiles/Splash.dir/playNote.cpp.o" \
-"CMakeFiles/Splash.dir/processBuffer.cpp.o" \
 "CMakeFiles/Splash.dir/realtime-beat.cpp.o" \
 "CMakeFiles/Splash.dir/splash.cpp.o"
 
@@ -184,7 +170,6 @@ bin/Splash: CMakeFiles/Splash.dir/RtAudio.cpp.o
 bin/Splash: CMakeFiles/Splash.dir/fft.cpp.o
 bin/Splash: CMakeFiles/Splash.dir/instructionsStatements.cpp.o
 bin/Splash: CMakeFiles/Splash.dir/playNote.cpp.o
-bin/Splash: CMakeFiles/Splash.dir/processBuffer.cpp.o
 bin/Splash: CMakeFiles/Splash.dir/realtime-beat.cpp.o
 bin/Splash: CMakeFiles/Splash.dir/splash.cpp.o
 bin/Splash: CMakeFiles/Splash.dir/build.make
@@ -198,7 +183,7 @@ bin/Splash: /usr/lib/arm-linux-gnueabihf/libsfml-graphics.so
 bin/Splash: /usr/lib/arm-linux-gnueabihf/libsfml-network.so
 bin/Splash: /usr/lib/arm-linux-gnueabihf/libsfml-audio.so
 bin/Splash: CMakeFiles/Splash.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_8) "Linking CXX executable bin/Splash"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/home/pi/sound-pirates/Urho3D/splash_build/CMakeFiles --progress-num=$(CMAKE_PROGRESS_7) "Linking CXX executable bin/Splash"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/Splash.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
