+++ Urho3D/splash_build/CMakeFiles/Splash.dir/link.txt
@@ -1,1 +1,1 @@
-/usr/bin/c++  -Wno-invalid-offsetof -pipe -mfloat-abi=hard -Wno-psabi -mcpu=cortex-a7 -mfpu=vfpv4-d16 -pthread -O3 -DNDEBUG -rdynamic CMakeFiles/Splash.dir/RtAudio.cpp.o CMakeFiles/Splash.dir/fft.cpp.o CMakeFiles/Splash.dir/instructionsStatements.cpp.o CMakeFiles/Splash.dir/playNote.cpp.o CMakeFiles/Splash.dir/processBuffer.cpp.o CMakeFiles/Splash.dir/realtime-beat.cpp.o CMakeFiles/Splash.dir/splash.cpp.o -o bin/Splash  -Wl,-rpath,/opt/vc/lib /opt/vc/lib/libbcm_host.so -lEGL -lGLESv2 /home/pi/Urho3D/lib/libUrho3D.a -ldl -lrt -lsfml-system -lsfml-window -lsfml-graphics -lsfml-network -lsfml-audio -lrtaudio -lfftw3 -lasound -lpthread -ljack 
+/usr/bin/c++  -Wno-invalid-offsetof -pipe -mfloat-abi=hard -Wno-psabi -mcpu=cortex-a7 -mfpu=vfpv4-d16 -pthread -O3 -DNDEBUG -rdynamic CMakeFiles/Splash.dir/RtAudio.cpp.o CMakeFiles/Splash.dir/fft.cpp.o CMakeFiles/Splash.dir/instructionsStatements.cpp.o CMakeFiles/Splash.dir/playNote.cpp.o CMakeFiles/Splash.dir/realtime-beat.cpp.o CMakeFiles/Splash.dir/splash.cpp.o -o bin/Splash  -Wl,-rpath,/opt/vc/lib /opt/vc/lib/libbcm_host.so -lEGL -lGLESv2 /home/pi/Urho3D/lib/libUrho3D.a -ldl -lrt -lsfml-system -lsfml-window -lsfml-graphics -lsfml-network -lsfml-audio -lrtaudio -lfftw3 -lasound -lpthread -ljack
