+++ Urho3D/splash_build/CMakeFiles/progress.marks
@@ -1,1 +1,1 @@
-8
+7
